pyproject.toml
[project]
name = "agentcore-examples"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "strands-agents[litellm]>=1.0.0",
    "strands-agents-tools>=0.2.1",
    "uv>=0.7.21",
    "boto3>=1.39.7",
    "bedrock-agentcore-starter-toolkit>=0.1.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "pyjwt>=2.9.0",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.8",
    "pytest>=8.3.0",
]
